requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
pyjwt==1.7.1

flake8==3.9.1